Limit method_missing to only handle known commands
Redis-Namespace has had a *long* history of issues rising from the blind
passing of unknown commands to the underlying Redis Client (see almost
every single issue in the [tracker][]). This patch is a breaking change
that stops that behaviour, instead raising a NoMethodError when handling
a command that is unknown.

TODO before 2.x release:
 - [ ] Add support & specs for Command Aliases
 - [ ] Define all methods to avoid method_missing altogether
       (via `Module#define_method`?)

[tracker]: https://github.com/resque/redis-namespace/issues?state=closed

Pairing session with @chelseakomlo and @yaauie.