Fix the travis build for rubinius :grimacing:
Rubinius no longer supplies 1.8-mode or 1.9-mode, and they have changed
how modes are to be selected entirely, along with the 2.x breaking change
to extract the ruby standard library into `rubysl-*` gems.

Effectively this means that we can no longer verify support for 1.8 mode
or 1.9 mode, and that we must add the rubysl gems as platform-specific
dependencies to the Gemfile and require that consumers of redis-namespace
on rubinius also have these gems on their load path.

Le Sigh...