better fix for UTF-8 keys issue
This fixes the frozen string issue from before.

Ruby 1.8.7 doesn't support string encoding, so just skip this fanciness for that version.

Updates the tests to demonstrate setting a key with invalid UTF-8 chars, getting a list of the keys (including the invalid string) and retrieving the data at those keys.